ISSUE #274 * Refactor `MinosModel` tests and prepare failing case.
