+++ tests/test_common/test_model/test_abc/test_model/__init__.py
@@ -1,0 +1,7 @@
+"""
+Copyright (C) 2021 Clariteia SL
+
+This file is part of minos framework.
+
+Minos framework can not be copied and/or distributed without the express permission of Clariteia SL.
+"""
